Updates to code have created errors Rolling back these changes to a point when the code was not broken in the hope that I can recover so functional code to deliver.
Script 1 and 2 updates. Making some minor changes to what is displayed to the user so that each step is clearly separated.

Creating functions to check if a shared folder exists, create if it does not. Debugged the fallback option in the event the local csv file is not located.

Debugging the parsing of files, the script now is stuck where it is checking the number of users that will be created.