Task 2 Completed (#8)
* Debugging the parsing of files, the script now is stuck where it is checking the number of users that will be created.

* more debugging

* Updated Task-1 Script

debugging the error occuring when parsing the locally stored csv file; unable to parse data from this

* Update to user create function, now can see that the environment that I am coding in is causing some issues.

* Updated download csv function so this should now parse user data

that is downloaded from a remote server

* Update to shared directory function

Just need to set permissions and change ownership of the shared folders

* task 2 script started. Also created the BSA text file required for the assignment documentation. Task 2 script has a menu script which tells the user what the script will do. confirms user wishes to continue.

* Task 1 code modified, trying to create the user shared directory

unsure where in the file structure this needs to be placed, still have to create task 2, as well as finish task 1. Task 1 still needs a alias placed in the user home directory + a softlink to the shared directory

* Update to assignment

* Starting task 2 - added a function for the user to enter a directory to compress, given a fallback option for the user to just use the current directory.

* user file now getting zipped

* updated task 2 - working on function to send zip file to remote host.

* Updates to task one added. Working through how to create a shared file with correct permissions

* Update to task one

Moved functionality from parseData function to streamline creation of user groups and shared folders

* Update to task 1 script. Debugging user shared folder create functions

* Update to task 1 script. Debugging user shared folder create functions

* Update to user home directory function

* Task 1 code - testing debugged code

* Task 1 debugging code

* Update to task 1 code - debugging user shared file functions

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Issues with the shut down alias function.

* Update to task 1 script - debugging code is hopefully completed.

Issues with the shut down alias function.

* Update to task 1 script - debugging code is hopefully completed.

Ongoing issues with the alias function, working through how to create the user, with a bash alias file in their home directory. for some reason this is not happening currently

* Update to task 1 script - debugging code is hopefully completed.

Ongoing issues with the alias function, working through how to create the user, with a bash alias file in their home directory. for some reason this is not happening currently

* Update to task 1 script - debugging code is hopefully completed.

Issues with the shut down alias function.

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - debugging code is hopefully completed.

Will test on VM next

* Update to task 1 script - have debugged the user password function.

This has been tested and should now be working  also need to fix the alias which is still not working correctly

* Update to task 1 script - working to debug the alias function

this is still have issues with correct permissions, also need to change the owner ship of the bash alias file to the user

* Update to task 1 script - working to debug the alias function

Have debugged the alias function - now testing

* Update to task 1 script - working on debugging the file download script this is having issues

* Updated task 1 code - Ready for testing

* Updated task 1, debugging issue with download function

New code added for testing

* Updated task 1 - download function now works

Now fixing issue with user creation and ownership assignment for the shared folders. Have added some changes for testing in the vm

* Update to task 1 script. Debugging file download script which is still not functioning correctly

* Update to task 1 script. Debugging file download script which is still not functioning correctly

* Update to task 1 script. Debugging file download script which is still not functioning correctly

* Debugging a new issue with the function

Needed to provide a fall back group and shared folder if one is not provided

* Debugging a new issue with the function

Needed to provide a fall back group and shared folder if one is not provided

* Update to task 1 script. Debugging file download script which is still not functioning correctly

* Was unable to resolve the issue of missing parameters

when the script runs an error occurs if there is no group or no shared folder supplied.

* debugging out logging

* debugging out logging

* Still debugging shared permissions

* debugging out logging

* Updated user scipt task 1

* Update to task-2 - added some functions that need to be completed

for the task. Just need to add code to these, plus error checking

* Updated both task scripts

Task 1, altered the delay function in some places as these had been doubled up, causing long pauses between functions. Also added a loops to check the user input directory in task 2. Commiting these for testing in the VM

* Update to task 2

* Update to task 2

* Update to task 2 script

Committing to test on the VM

* Update task 2 assignment

* Updated task 2 scripts, debugged inital error checking loops that give fallback options

in the event the user enters incorrect information

* Updated gitignore, deleted log.txt for task.2

* Update to Task 2 Code

Still incomplete, working through getting user input data and error checking it

* Updates to task 2

Added better error check and loops to prevent infinite errors

* Update to README.md

* Update to README.md

* Update to README.md

* Update to task 2 - script is now error checking input ip and user name is provided

* Update to task 2 - script is now error checking input ip and user name is provided

* Updates to task 2 script. Debugging the error checking in the IP address function

* Debugging the zip file function which has stopped working on my development computer

* Task 2 updated, script is now at the point of transferring zip to remote

* Updates to task 2 script - now ready for testing on VMs

* Task 2 code tested, is now completed and ready to merge to main branch for assessment

* directory clean up for assignment submission

Co-authored-by: Anthony Legg <>